# match/infra/api/task.py
from dataclasses import asdict
from http import HTTPStatus

from fastapi import APIRouter, Depends, File, HTTPException, Request, Response, UploadFile

from match.app.service import MatchService
from match.bootstrap import get_service
from match.domain.exceptions import (
    DomainException,
    ImageNotFound,
    InvalidTaskAction,
    MatchServiceException,
    PermissionDenied,
    TaskNotFound,
)
from match.domain.interfaces import TaskFilter
from match.domain.task import Category, LocationRadius, Task, TaskStatus
from match.domain.user import User
from match.infra.api.auth import get_user_id, verified_user
from match.infra.api.schemas import (
    PublicTaskSchema,
    TaskAction,
    TaskCreationRequestSchema,
    TaskEditRequestSchema,
    TaskLocationSchema,
    TaskSchema,
)

router = APIRouter()


def _task_filters_from_request(request: Request) -> TaskFilter:
    query_params = request.query_params
    filters: TaskFilter = {}
    location_filter_keys = {"lat", "lon", "radius_km"}

    try:
        if "status" in query_params:
            filters["status"] = TaskStatus(query_params["status"])
        if "category" in query_params:
            filters["category"] = Category(query_params["category"])
        if "owner_id" in query_params:
            filters["owner_id"] = int(query_params["owner_id"])
        if "helper_id" in query_params:
            helper_id = query_params["helper_id"]
            filters["helper_id"] = None if helper_id == "null" else int(helper_id)
        if location_filter_keys & query_params.keys():
            if not location_filter_keys <= query_params.keys():
                raise ValueError
            location_radius = LocationRadius(
                lat=float(query_params["lat"]),
                lon=float(query_params["lon"]),
                radius_km=float(query_params["radius_km"]),
            )
            if not -90 <= location_radius.lat <= 90:
                raise ValueError
            if not -180 <= location_radius.lon <= 180:
                raise ValueError
            if location_radius.radius_km <= 0:
                raise ValueError
            filters["location_radius"] = location_radius
    except ValueError as exc:
        raise HTTPException(
            status_code=HTTPStatus.BAD_REQUEST, detail="Invalid task filter."
        ) from exc

    return filters


def public_task_to_dict(task: Task) -> dict:
    task_dict = asdict(task)
    task_dict.pop("owner_id")
    task_dict.pop("helper_id")
    task_dict.pop("created_at")
    task_dict.pop("updated_at")
    task_dict.pop("description")
    return task_dict


@router.post("/", response_model=TaskSchema, status_code=HTTPStatus.CREATED)
def create_task(
    task_creation_params: TaskCreationRequestSchema,
    user: User = Depends(verified_user),
    service: MatchService = Depends(get_service),
) -> dict:
    try:
        task = service.create_task(
            user.id,
            description=task_creation_params.description,
            title=task_creation_params.title,
            category=task_creation_params.category,
            location_lon=task_creation_params.location.lon,
            location_lat=task_creation_params.location.lat,
            location_address=task_creation_params.location.address,
        )
        return service.format_task_response(task)
    except PermissionDenied:
        raise HTTPException(status_code=HTTPStatus.FORBIDDEN)


@router.get("/images/{image_id}")
def get_task_image(image_id: str, service: MatchService = Depends(get_service)) -> Response:
    try:
        content = service.get_task_image(image_id)
    except (ImageNotFound, FileNotFoundError):
        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
    return Response(content=content, media_type="application/octet-stream")


@router.post("/{task_id}/images", response_model=TaskSchema, status_code=HTTPStatus.CREATED)
def add_task_images(
    task_id: int,
    images: list[UploadFile] = File(...),
    user: User = Depends(verified_user),
    service: MatchService = Depends(get_service),
) -> dict:
    try:
        task = service.task_add_images(
            task_id, owner_id=user.id, images=[image.file.read() for image in images]
        )
    except TaskNotFound:
        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
    except InvalidTaskAction as e:
        raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail=str(e))
    return service.format_task_response(task)


@router.delete("/{task_id}/images/{image_id}", response_model=TaskSchema)
def remove_task_image(
    task_id: int,
    image_id: str,
    user: User = Depends(verified_user),
    service: MatchService = Depends(get_service),
) -> dict:
    try:
        task = service.task_remove_image(task_id, owner_id=user.id, image_id=image_id)
    except TaskNotFound:
        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
    except InvalidTaskAction as e:
        raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail=str(e))
    except DomainException:
        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
    return service.format_task_response(task)


@router.put("/{task_id}/edit", response_model=TaskSchema)
def edit_task(
    task_id: int,
    task_edit_params: TaskEditRequestSchema,
    user: User = Depends(verified_user),
    service: MatchService = Depends(get_service),
) -> dict:
    location = task_edit_params.location
    try:
        task = service.task_edit(
            task_id,
            owner_id=user.id,
            title=task_edit_params.title,
            description=task_edit_params.description,
            category=task_edit_params.category,
            location_lon=location.lon if location else None,
            location_lat=location.lat if location else None,
            location_address=location.address if location else None,
        )
    except TaskNotFound:
        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)
    except InvalidTaskAction as e:
        raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail=str(e))
    except MatchServiceException:
        raise HTTPException(status_code=HTTPStatus.BAD_REQUEST)

    return service.format_task_response(task)


@router.put("/{task_id}/manage", response_model=TaskSchema)
def manage_task(
    task_id: int,
    action: TaskAction,
    message: str | None = None,
    helper_id: int | None = None,
    user: User = Depends(verified_user),
    service: MatchService = Depends(get_service),
) -> dict:
    try:
        match action:
            case TaskAction.JOIN:
                if message is None:
                    raise HTTPException(
                        status_code=HTTPStatus.BAD_REQUEST,
                        detail="Message is required to join a task.",
                    )
                task = service.task_join(task_id, user.id, message)
            case TaskAction.APPROVE:
                if helper_id is None:
                    raise HTTPException(
                        status_code=HTTPStatus.BAD_REQUEST, detail="Helper id not provided."
                    )
                task = service.task_approve(task_id, owner_id=user.id, helper_id=helper_id)
            case TaskAction.REJECT:
                if helper_id is None:
                    raise HTTPException(
                        status_code=HTTPStatus.BAD_REQUEST, detail="Helper id not provided."
                    )
                task = service.task_reject(task_id, owner_id=user.id, helper_id=helper_id)
            case TaskAction.CLOSE:
                task = service.task_close(task_id, owner_id=user.id)
            case TaskAction.REPORT_SUCCESS:
                task = service.task_report_success(task_id, owner_id=user.id)
            case TaskAction.REPORT_FAILURE:
                task = service.task_report_failed(task_id, owner_id=user.id)
    except PermissionDenied as e:
        raise HTTPException(
            status_code=HTTPStatus.FORBIDDEN,
            detail=f"Permission denied for user {user}.{str(e)}",
        )

    return service.format_task_response(task)


@router.get("/", response_model=list[TaskSchema])
def list_tasks(
    request: Request,
    service: MatchService = Depends(get_service),
    _: User = Depends(verified_user),
) -> list:
    return service.get_tasks_response(filters=_task_filters_from_request(request))


@router.get("/locations", response_model=list[TaskLocationSchema])
def list_task_locations(
    request: Request,
    service: MatchService = Depends(get_service),
) -> list:
    return service.get_task_locations(filters=_task_filters_from_request(request))


@router.get("/public", response_model=list[PublicTaskSchema])
def list_tasks_public(
    request: Request,
    service: MatchService = Depends(get_service),
) -> list:
    return service.get_tasks_response(filters=_task_filters_from_request(request))


@router.get("/my-tasks", response_model=list[TaskSchema])
def get_my_tasks(
    user_id: int = Depends(get_user_id), service: MatchService = Depends(get_service)
) -> list[dict]:
    return service.get_tasks_response(filters={"owner_id": user_id})


@router.get("/{task_id}", response_model=TaskSchema)
def get_task(
    task_id: int, _: User = Depends(verified_user), service: MatchService = Depends(get_service)
) -> dict:
    try:
        return service.get_task_response(task_id)
    except TaskNotFound:
        raise HTTPException(status_code=HTTPStatus.NOT_FOUND)