# openapi_specs.json
{
  "components": {
    "schemas": {
      "Body_add_task_images_task__task_id__images_post": {
        "properties": {
          "images": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Images",
            "type": "array"
          }
        },
        "required": [
          "images"
        ],
        "title": "Body_add_task_images_task__task_id__images_post",
        "type": "object"
      },
...
      "Category": {
        "enum": [
          "transport people",
          "food",
          "accommodation",
          "clothes",
          "medical help",
          "clean",
          "repair",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageSchema": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "path": {
            "format": "uri",
            "minLength": 1,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path"
        ],
        "title": "ImageSchema",
        "type": "object"
      },
      "Location": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Lon"
          }
        },
        "required": [
          "lat",
          "lon",
          "address"
        ],
        "title": "Location",
        "type": "object"
      },
...
      "TaskAction": {
        "enum": [
          "join",
          "approve",
          "reject",
          "close",
          "report_success",
          "report_failure"
        ],
        "title": "TaskAction",
        "type": "string"
      },
      "TaskCreationRequestSchema": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "category",
          "location"
        ],
        "title": "TaskCreationRequestSchema",
        "type": "object"
      },
      "TaskEditRequestSchema": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description",
          "category",
          "location"
        ],
        "title": "TaskEditRequestSchema",
        "type": "object"
      },
      "TaskEventSchema": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/TaskUserSchema"
          },
          "helper": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskUserSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TaskEventType"
          }
        },
        "required": [
          "id",
          "type",
          "actor",
          "helper",
          "message",
          "occurred_at"
        ],
        "title": "TaskEventSchema",
        "type": "object"
      },
      "TaskEventType": {
        "enum": [
          "created",
          "offered",
          "approved",
          "rejected",
          "closed",
          "succeeded",
          "failed"
        ],
        "title": "TaskEventType",
        "type": "string"
      },
...
      "TaskSchema": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TaskEventSchema"
            },
            "title": "Events",
            "type": "array"
          },
          "helper": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskUserSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "helper_offers": {
            "items": {},
            "title": "Helper Offers",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageSchema"
            },
            "title": "Images",
            "type": "array"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "owner": {
            "$ref": "#/components/schemas/TaskUserSchema"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at",
          "status",
          "owner",
          "helper",
          "helper_offers",
          "description",
          "location",
          "category",
          "images",
          "events"
        ],
        "title": "TaskSchema",
        "type": "object"
      },
...
      "TaskStatus": {
        "enum": [
          "open",
          "pending",
          "approved",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "TaskUserSchema": {
        "properties": {
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "first_name"
        ],
        "title": "TaskUserSchema",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "user/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "match",
    "version": "0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/task/": {
      "get": {
        "operationId": "list_tasks_task__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSchema"
                  },
                  "title": "Response List Tasks Task  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Tasks",
        "tags": [
          "task"
        ]
      },
      "post": {
        "operationId": "create_task_task__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreationRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Task",
        "tags": [
          "task"
        ]
      }
    },
    "/task/images/{image_id}": {
      "get": {
        "operationId": "get_task_image_task_images__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Image",
        "tags": [
          "task"
        ]
      }
    },
...
    "/task/my-tasks": {
      "get": {
        "operationId": "get_my_tasks_task_my_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskSchema"
                  },
                  "title": "Response Get My Tasks Task My Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get My Tasks",
        "tags": [
          "task"
        ]
      }
    },
...
    "/task/{task_id}": {
      "get": {
        "operationId": "get_task_task__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Task",
        "tags": [
          "task"
        ]
      }
    },
    "/task/{task_id}/edit": {
      "put": {
        "operationId": "edit_task_task__task_id__edit_put",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskEditRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Edit Task",
        "tags": [
          "task"
        ]
      }
    },
    "/task/{task_id}/images": {
      "post": {
        "operationId": "add_task_images_task__task_id__images_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_task_images_task__task_id__images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Task Images",
        "tags": [
          "task"
        ]
      }
    },
    "/task/{task_id}/images/{image_id}": {
      "delete": {
        "operationId": "remove_task_image_task__task_id__images__image_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove Task Image",
        "tags": [
          "task"
        ]
      }
    },
    "/task/{task_id}/manage": {
      "put": {
        "operationId": "manage_task_task__task_id__manage_put",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TaskAction"
            }
          },
          {
            "in": "query",
            "name": "message",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Message"
            }
          },
          {
            "in": "query",
            "name": "helper_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Helper Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Manage Task",
        "tags": [
          "task"
        ]
      }
    },
...
  }
}